Fix conversion for String type; refactor related functions to Arrow.scala
