+++ sql/core/src/test/resources/test-data/arrowNullInts.json
@@ -15,6 +15,7 @@
             }
         ]
     },
+
     "batches": [
         {
             "count": 4,
