+++ sql/core/src/test/scala/org/apache/spark/sql/ArrowSuite.scala
@@ -19,14 +19,13 @@ package org.apache.spark.sql
 import java.io.File
 
 import org.apache.arrow.memory.RootAllocator
-import org.apache.arrow.vector.{BitVector, VectorLoader, VectorSchemaRoot}
+import org.apache.arrow.vector.{VectorLoader, VectorSchemaRoot}
 import org.apache.arrow.vector.file.json.JsonFileReader
 import org.apache.arrow.vector.util.Validator
 
-import org.apache.spark.sql.test.{SharedSQLContext, SQLTestUtils}
+import org.apache.spark.sql.test.{ArrowTestData, SharedSQLContext}
 
-class ArrowSuite extends QueryTest with SharedSQLContext with SQLTestUtils {
-  private val nullIntsFile = "test-data/arrowNullInts.json"
+class ArrowSuite extends QueryTest with SharedSQLContext with ArrowTestData {
 
   private def testFile(fileName: String): String = {
     // TODO: Copied from CSVSuite, find a better way to read test files
@@ -34,13 +33,20 @@ class ArrowSuite extends QueryTest with SharedSQLContext with SQLTestUtils {
   }
 
   test("convert int column with null to arrow") {
-    val df = nullInts
-    val jsonFilePath = testFile(nullIntsFile)
+    test(arrowNullInts, arrowNullIntsFile)
+  }
+
+  test("convert string column with null to arrow") {
+    test(arrowNullStrings, arrowNullStringsFile)
+  }
+
+  private def test(df: DataFrame, arrowFile: String) {
+    val jsonFilePath = testFile(arrowFile)
 
     val allocator = new RootAllocator(Integer.MAX_VALUE)
     val jsonReader = new JsonFileReader(new File(jsonFilePath), allocator)
 
-    val arrowSchema = df.schemaToArrowSchema(df.schema)
+    val arrowSchema = Arrow.schemaToArrowSchema(df.schema)
     val jsonSchema = jsonReader.start()
     Validator.compareSchemas(arrowSchema, jsonSchema)
